Regularize dr_logger.py internal code, particularly registry format.
The new registry format is described in the docstring for
DRLoggerServer.

The server will now start up properly if one or more of the servers
it watches is not available at startup. The datavault dataset
created has as many columns as available variables to measure. In
other words, if some servers are missing we just make a dataset with
appropriately fewer columns. This is nice but the main point of the
changes here are to clean up the code so that future work can make
the server more rhobust against intermittent failures of other
resources on the LabRAD system.

Review: ChristopherQuintana