Remove import_parameters setting on data_vault server.
This setting imports a bunch of parameters from the registry by using
the context of the given request and assuming that the same context
has been set up with the registry. This sort of inter-server context
synchronization is a bad idea and needs to go away. This setting is not
actually being used, instead we usually just set parameters on our
own using the registry wrapper, which is a much more sane way to go.

Review: ejeffrey