Change ghz_fpga_server to always read a list of packets.
The delphi direct ethernet server changes the 'shape' of packet responses
when you try to read one versus multiple packets. In particular, if you
call .read() it returns a single packet as (ssis), but if you call .read(1)
it returns a list of one packet as *(ssis), which obviously is the same
type as when the number of packets to be read is larger. By changing all
the read calls to specify the number of packets explicitly, we always
get the same type back, which normalizes the response handling. In the
new scala version of the direct ethernet server, this allows us to simplify
the implementation since we can just specify a single return type.

Review: @DanielSank, @ejeffrey