Fix a few bugs with CYCLE handling.
There is a fix in the qubit sequencer as well as
an offset to jump table index in the fpga server.

This compensates for the first entry of the jump table
always being a NOP unspecified by the user. Because of this, the user's
first entry is actually 1. So, we add 1.

Bump version to 0.6.1.

Review: maffoo.