datavault: Fix labrad_urlencode to work with new FlatData type
The FlatData type is a namedtuple that replaces the plain tuple
returned by older versions of pylabrad when flattening data. We
modify the labrad_urlencode method to use names of the FlatData
items if possible, rather than explicitly unpacking the tuple.
This lets us maintain compatibility with changes in 0.95.2 which
added a third endianness member to the FlatData tuple.